requirements.txt
streamlit>=1.30.0
vaderSentiment>=3.3.2
numpy>=1.26.0
sounddevice>=0.4.6
openai-whisper>=20231117
torch>=2.2.0
huggingface-hub>=0.24.0
sentencepiece>=0.2.0
